Refactored
Changed dimension keys to strongly typed values
Added extension methods to get the next and previous key
Rider Code refactor

+++ DateTimeProviders.Providers/Extensions/JavaTimeStampExtensions.cs
@@ -3,7 +3,7 @@
 public static class JavaTimeStampExtensions
 {
     /// <summary>
-    /// Convert Java Timestamp to datetime.
+    ///     Convert Java Timestamp to datetime.
     /// </summary>
     /// <param name="javaTimeStamp">the Java TimeStamp.</param>
     /// <returns>a DateTime</returns>
@@ -11,17 +11,18 @@ public static class JavaTimeStampExtensions
     {
         // Unix timestamp is seconds past epoch
         var dtDateTime = new DateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeKind.Utc);
-        dtDateTime = dtDateTime.AddMilliseconds(javaTimeStamp*1000).ToLocalTime();
+        dtDateTime = dtDateTime.AddMilliseconds(javaTimeStamp * 1000).ToLocalTime();
         return dtDateTime;
     }
 
     /// <summary>
-    /// Convert DateTime to Java Timestamp
+    ///     Convert DateTime to Java Timestamp
     /// </summary>
     /// <param name="dateTime">the DateTime</param>
     /// <returns>a Java Timestamp</returns>
     public static double ToJavaTimestamp(this DateTime dateTime)
     {
-        return (double)(dateTime.ToUniversalTime() - new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc)).TotalMilliseconds/1000;
+        return (dateTime.ToUniversalTime() - new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc)).TotalMilliseconds /
+               1000;
     }
 }
